refactor: exported functions from services into correct files
