+++ src/network/mod.rs
@@ -2,3 +2,5 @@ pub mod forward;
 pub mod message;
 pub mod server;
 pub mod peer_ipc;
+pub mod peers_operations;
+pub mod watchdogs;
